Minor Fix & Changes
Hi, at first, this project is a totally amazing work! Damn fast and reliable.

We made a few adjustments and wanted to contribte them. Maybe this may seem useful for you too.
- Fix: nfcursortoobject() throws error when cursor has a field "aname" in it.
- Change: added new parameter "cAlias" to nfcursortojson4vfp() and nfcursortojson(). Since nfJsonToCursor() also has a parameter to name the alias this seems  conclusive.
- Added: nfJsonBeautify() so on can easy make an existing JSON-String readable/formattet.

Hope u like it.
Thanks a lot for this great project (again) and keep up the good work. :)